format and comments, spring cleaning
