+++ src/cli.c
@@ -1,32 +1,40 @@
 // File: cil.c
 // -----------------------
-#include <stdarg.h>   // for va_list, va_start, va_end
-#include <stdbool.h>  // for bool, true, false
-#include <stddef.h>   // for NULL
-#include <stdio.h>    // for printf, vprintf, fgets, stdin
-#include <string.h>   // for strcspn
+// Implements basic command-line interface (CLI) utilities.
+
+#include <stdarg.h>   // va_list, va_start, va_end
+#include <stdbool.h>  // bool, true, false
+#include <stddef.h>   // NULL
+#include <stdio.h>    // printf, vprintf, fgets, stdin
+#include <string.h>   // strcspn
+
 // Local includes
 #include "cli.h"
+
 /**
- * Function to read a line from the command line
+ * Reads a line from the standard input into the provided buffer.
+ * Trims the newline character if present.
  *
- * @param buf the buffer to read into
- * @return true if a line was read, false otherwise
+ * @param buf A buffer of at least CLI_LINESZ to store the input.
+ * @return true if input was successfully read, false otherwise.
  */
 bool cli_readline(char buf[static CLI_LINESZ]) {
-    if (NULL == fgets(buf, CLI_LINESZ, stdin))
+    if (fgets(buf, CLI_LINESZ, stdin) == NULL)
         return false;
+    
+    // Trim newline character if present
     buf[strcspn(buf, "\n")] = '\0';
     return true;
 }
+
 /**
- * Function to print a line to the command line
+ * Prints a formatted line to the command line, prefixed with '>'.
  *
- * @param format the format string
- * @param ... the arguments to format
+ * @param format Format string (similar to printf).
+ * @param ... Variadic arguments matching the format specifier.
  */
 void cli_println(const char *format, ...) {
-    printf(">");
+    printf(">");  // CLI prompt prefix
 
     va_list args;
     va_start(args, format);
@@ -35,4 +43,3 @@ void cli_println(const char *format, ...) {
 
     printf("\n");
 }
-
