+++ src/ui.h
@@ -1,0 +1,2 @@
+void show_notification(WINDOW *win, const char *format, ...);
+void show_popup(const char *title, const char *fmt, ...);
