+++ src/utils.c
@@ -1,5 +1,7 @@
 // File: utils.c
 // -----------------------
+#define _GNU_SOURCE
+#define _POSIX_C_SOURCE 200809L
 #include <errno.h>     // for errno
 #include <stdarg.h>    // for va_list, va_start, va_end
 #include <stdio.h>     // for fprintf, stderr, vfprintf
@@ -13,7 +15,9 @@
 #include <sys/stat.h>  // for stat, S_ISDIR
 #include <ctype.h>     // for isprint
 #include <libgen.h>    // for dirname() and basename()
-
+#include <time.h>      // for clock_gettime
+#include <unistd.h>    // Required for usleep()
+#include <magic.h> 
 // Local includes
 #include "utils.h"
 #include "files.h"  // Include the header for FileAttr and related functions
@@ -814,4 +818,319 @@ void reload_directory(Vector *files, const char *current_directory) {
     append_files_to_vec(files, current_directory);
     // Makes the vector shorter
     Vector_sane_cap(files);
+}
+
+
+/**
+ * Function to draw and scroll the banner text
+ *
+ * @param window the banner window
+ * @param text the text to scroll
+ * @param build_info the build information to display
+ * @param offset the current offset for scrolling
+ */
+void draw_scrolling_banner(WINDOW *window, const char *text, const char *build_info, int offset) {
+    struct timespec current_time;
+    clock_gettime(CLOCK_MONOTONIC, &current_time);
+
+    // Only update banner if enough time has passed
+    long time_diff = (current_time.tv_sec - last_scroll_time.tv_sec) * 1000000 +
+                     (current_time.tv_nsec - last_scroll_time.tv_nsec) / 1000;
+
+    if (time_diff < BANNER_SCROLL_INTERVAL) {
+        return;  // Skip update if not enough time has passed
+    }
+
+    // Fetch current local time
+    time_t rawtime;
+    struct tm *timeinfo;
+    char time_str[9];  // HH:MM:SS format
+
+    time(&rawtime);
+    timeinfo = localtime(&rawtime);
+    strftime(time_str, sizeof(time_str), "%H:%M:%S", timeinfo);
+
+    int width = COLS - 2;
+    int text_len = strlen(text);
+    int build_len = strlen(build_info);
+    int time_len = strlen(time_str);
+
+    // Calculate total length including padding
+    int total_len = width + text_len + build_len + time_len + 6;
+
+    // Create the scroll text buffer
+    char *scroll_text = malloc(2 * total_len + 1);
+    if (!scroll_text) return;
+
+    memset(scroll_text, ' ', 2 * total_len);
+    scroll_text[2 * total_len] = '\0';
+
+    // Copy the text pattern twice for smooth wrapping
+    for (int i = 0; i < 2; i++) {
+        int pos = i * total_len;
+        snprintf(scroll_text + pos, total_len, "%s | %s | %s", text, build_info, time_str);
+    }
+
+    // Draw the banner
+    werase(window);
+    box(window, 0, 0);
+    mvwprintw(window, 1, 1, "%.*s", width, scroll_text + offset);
+    wrefresh(window);
+
+    free(scroll_text);
+
+    // Update last scroll time
+    last_scroll_time = current_time;
+}
+
+// Function to handle banner scrolling in a separate thread
+void *banner_scrolling_thread(void *arg) {
+    WINDOW *window = (WINDOW *)arg;
+    int banner_offset = 0;
+    struct timespec last_update_time;
+    clock_gettime(CLOCK_MONOTONIC, &last_update_time);
+
+    int total_scroll_length = COLS + strlen(BANNER_TEXT) + strlen(BUILD_INFO) + 4;
+
+    while (1) {
+        struct timespec current_time;
+        clock_gettime(CLOCK_MONOTONIC, &current_time);
+        long time_diff = (current_time.tv_sec - last_update_time.tv_sec) * 1000000 +
+                         (current_time.tv_nsec - last_update_time.tv_nsec) / 1000;
+
+        if (time_diff >= BANNER_SCROLL_INTERVAL) {
+            draw_scrolling_banner(window, BANNER_TEXT, BUILD_INFO, banner_offset);
+            banner_offset = (banner_offset + 1) % total_scroll_length;
+            last_update_time = current_time;
+        }
+
+        // Sleep for a short duration to prevent busy-waiting
+        usleep(10000); // 10ms
+    }
+
+    return NULL;
+}
+
+/** Function to cleanup tmp files made by operations
+*/
+void cleanup_temp_files() {
+    char command[1024];
+    snprintf(command, sizeof(command), "rm -rf /tmp/cupidfm_*_%d", getpid());
+    system(command);
+}
+
+const char* keycode_to_string(int keycode) {
+    static char buf[32];
+
+    // Define the base value for function keys
+    // Typically, KEY_F(1) is 265, so base = 264
+    const int FUNCTION_KEY_BASE = KEY_F(1) - 1;
+
+    // Handle function keys
+    if (keycode >= KEY_F(1) && keycode <= KEY_F(63)) {
+        int func_num = keycode - FUNCTION_KEY_BASE;
+        snprintf(buf, sizeof(buf), "F%d", func_num);
+        return buf;
+    }
+
+    // Handle control characters (Ctrl+A to Ctrl+Z)
+    if (keycode >= 1 && keycode <= 26) { // Ctrl+A (1) to Ctrl+Z (26)
+        char c = 'A' + (keycode - 1);
+        snprintf(buf, sizeof(buf), "^%c", c);
+        return buf;
+    }
+
+    // Handle special keys
+    switch (keycode) {
+        case KEY_UP: return "KEY_UP";
+        case KEY_DOWN: return "KEY_DOWN";
+        case KEY_LEFT: return "KEY_LEFT";
+        case KEY_RIGHT: return "KEY_RIGHT";
+        case '\t': return "Tab";
+        case KEY_BACKSPACE: return "Backspace";
+        // Add more special keys as needed
+        default:
+            // Handle printable characters
+            if (keycode >= 32 && keycode <= 126) { // Printable ASCII
+                snprintf(buf, sizeof(buf), "%c", keycode);
+                return buf;
+            }
+            return "UNKNOWN";
+    }
+}
+
+/**
+ * Function to format a file size in a human-readable format
+ *
+ * @param buffer the buffer to store the formatted file size
+ * @param size the size of the file
+ * @return the formatted file size
+ */
+char* format_file_size(char *buffer, size_t size) {
+    // iB for multiples of 1024, B for multiples of 1000
+    // so, KiB = 1024, KB = 1000
+    const char *units[] = {"B", "KiB", "MiB", "GiB", "TiB"};
+    int i = 0;
+    double fileSize = (double)size;
+    while (fileSize >= 1024 && i < 4) {
+        fileSize /= 1024;
+        i++;
+    }
+    sprintf(buffer, "%.2f %s", fileSize, units[i]);
+    return buffer;
+}
+
+/** Function to show directory tree recursively
+ *
+ * @param window the window to display the directory tree
+ * @param dir_path the path of the directory to display
+ * @param level the current level of the directory tree
+ * @param line_num the current line number in the window
+ * @param max_y the maximum number of lines in the window
+ * @param max_x the maximum number of columns in the window
+ */
+void show_directory_tree(WINDOW *window, const char *dir_path, int level, int *line_num, int max_y, int max_x) {
+    if (level == 0) {
+        mvwprintw(window, 6, 2, "Directory Tree Preview:");
+        (*line_num)++;
+    }
+
+    // Early exit if we're already past visible area
+    if (*line_num >= max_y - 1) {
+        return;
+    }
+
+    DIR *dir = opendir(dir_path);
+    if (!dir) return;
+
+    struct dirent *entry;
+    struct stat statbuf;
+    char full_path[MAX_PATH_LENGTH];
+    size_t dir_path_len = strlen(dir_path);
+
+    // Define window size for entries
+    const int WINDOW_SIZE = 50; // Maximum entries to process at once
+    const int VISIBLE_ENTRIES = max_y - *line_num - 1; // Available lines in window
+    const int MAX_ENTRIES = MIN(WINDOW_SIZE, VISIBLE_ENTRIES);
+
+    struct {
+        char name[MAX_PATH_LENGTH];
+        bool is_dir;
+        mode_t mode;
+    } entries[WINDOW_SIZE];
+    int entry_count = 0;
+
+    // Only collect entries that will be visible
+    while ((entry = readdir(dir)) != NULL && entry_count < MAX_ENTRIES) {
+        if (strcmp(entry->d_name, ".") == 0 || strcmp(entry->d_name, "..") == 0) continue;
+
+        size_t name_len = strlen(entry->d_name);
+        if (dir_path_len + name_len + 2 > MAX_PATH_LENGTH) continue;
+
+        strcpy(full_path, dir_path);
+        if (full_path[dir_path_len - 1] != '/') {
+            strcat(full_path, "/");
+        }
+        strcat(full_path, entry->d_name);
+
+        if (lstat(full_path, &statbuf) == -1) continue;
+
+        // Only store if it will be visible
+        if (*line_num + entry_count < max_y - 1) {
+            strncpy(entries[entry_count].name, entry->d_name, MAX_PATH_LENGTH - 1);
+            entries[entry_count].name[MAX_PATH_LENGTH - 1] = '\0';
+            entries[entry_count].is_dir = S_ISDIR(statbuf.st_mode);
+            entries[entry_count].mode = statbuf.st_mode;
+            entry_count++;
+        }
+    }
+    closedir(dir);
+
+    // Check if no entries were found
+    if (entry_count == 0) {
+        mvwprintw(window, *line_num, 2 + level * 2, "This directory is empty");
+        (*line_num)++;
+    }
+
+    // Initialize magic only if we have entries to display
+    magic_t magic_cookie = NULL;
+    if (entry_count > 0) {
+        magic_cookie = magic_open(MAGIC_MIME_TYPE);
+        if (magic_cookie != NULL) {
+            magic_load(magic_cookie, NULL);
+        }
+    }
+
+    // Display collected entries
+    for (int i = 0; i < entry_count && *line_num < max_y - 1; i++) {
+        const char *emoji;
+        if (entries[i].is_dir) {
+            emoji = "📁";
+        } else if (magic_cookie) { // if magic fails 
+            size_t name_len = strlen(entries[i].name);
+            if (dir_path_len + name_len + 2 <= MAX_PATH_LENGTH) {
+                strcpy(full_path, dir_path);
+                if (full_path[dir_path_len - 1] != '/') {
+                    strcat(full_path, "/");
+                }
+                strcat(full_path, entries[i].name);
+                const char *mime_type = magic_file(magic_cookie, full_path);
+                emoji = get_file_emoji(mime_type, entries[i].name);
+            } else {
+                emoji = "📄";
+            }
+        } else {
+            emoji = "📄";
+        }
+
+        mvwprintw(window, *line_num, 2 + level * 2, "%s %.*s", 
+                  emoji, max_x - 4 - level * 2, entries[i].name);
+
+        char perm[10];
+        snprintf(perm, sizeof(perm), "%o", entries[i].mode & 0777);
+        mvwprintw(window, *line_num, max_x - 10, "%s", perm);
+        (*line_num)++;
+
+        // Only recurse into directories if we have space
+        if (entries[i].is_dir && *line_num < max_y - 1) {
+            size_t name_len = strlen(entries[i].name);
+            if (dir_path_len + name_len + 2 <= MAX_PATH_LENGTH) {
+                strcpy(full_path, dir_path);
+                if (full_path[dir_path_len - 1] != '/') {
+                    strcat(full_path, "/");
+                }
+                strcat(full_path, entries[i].name);
+                show_directory_tree(window, full_path, level + 1, line_num, max_y, max_x);
+            }
+        }
+    }
+
+    if (magic_cookie) {
+        magic_close(magic_cookie);
+    }
+}
+/** Function to return if file is hidden
+  * @return true or false
+ */
+bool is_hidden(const char *filename) {
+    return filename[0] == '.' && (strlen(filename) == 1 || (filename[1] != '.' && filename[1] != '\0'));
+}
+
+/** Function to get the total number of lines in a file
+ *
+ * @param file_path the path to the file
+ * @return the total number of lines in the file
+ */
+int get_total_lines(const char *file_path) {
+    FILE *file = fopen(file_path, "r");
+    if (!file) return 0;
+
+    int total_lines = 0;
+    char line[256];
+    while (fgets(line, sizeof(line), file)) {
+        total_lines++;
+    }
+
+    fclose(file);
+    return total_lines;
 }
